clean up some warnings dealing with 1) mismatches between printf format strings and arguments 2) wrong type for last parm to ap[r]_xml_to_text 3) missing time.h decls
git-svn-id: https://svn.apache.org/repos/asf/httpd/httpd/trunk@88603 13f79535-47bb-0310-9956-ffa450edef68

+++ modules/dav/main/util_lock.c
@@ -173,7 +173,7 @@ const char *dav_lock_get_activelock(request_rec *r, dav_lock *lock,
 	}
 	else {
 	    time_t now = time(NULL);
-	    sprintf(tmp, "Second-%lu", lock->timeout - now);
+	    sprintf(tmp, "Second-%lu", (long unsigned int)(lock->timeout - now));
 	    dav_buffer_append(p, pbuf, tmp);
 	}
